[PL] PF: Extract LocalAssemblerInterface.
